Converted UnitImpl::_getPlayer, UnitImpl::_getType, UnitImpl::_getPosition, and UnitImpl::_getTransport to variables.
Removed UnitImpl::getBWType() (UnitImpl::_getType works just as well).
Removed UnitImpl::staticTilePosition (now computed from staticPosition).
Removed UnitImpl::_getTilePosition (unnecessary).

Moved implementation of Unit::exists, Unit::getPlayer, Unit::getType, Unit::getPosition, and Unit::getTilePosition to UnitUpdate.cpp.

git-svn-id: https://bwapi.googlecode.com/svn/trunk@2481 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BWAPI/Server.cpp
@@ -474,7 +474,7 @@ namespace BWAPI
   }
   Force* Server::getForce(int id)
   {
-    if (forceVector.size()<=id)
+    if ((int)forceVector.size()<=id)
       return NULL;
     return forceVector[id];
   }
@@ -491,7 +491,7 @@ namespace BWAPI
   }
   Player* Server::getPlayer(int id)
   {
-    if (playerVector.size()<=id)
+    if ((int)playerVector.size()<=id)
       return NULL;
     return playerVector[id];
   }
@@ -508,7 +508,7 @@ namespace BWAPI
   }
   Unit* Server::getUnit(int id)
   {
-    if (unitVector.size()<=id)
+    if ((int)unitVector.size()<=id)
       return NULL;
     return unitVector[id];
   }
